% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trecUtils.R
\name{parseWebTopicDesc}
\alias{parseWebTopicDesc}
\title{Parse the TREC XML file with topics and subtopics}
\usage{
parseWebTopicDesc(filename)
}
\arguments{
\item{filename}{path to the TREC XML topics file}
}
\value{
a list with the topics and subtopics
}
\description{
Parse the TREC XML file with topics and subtopics
}
\examples{
# There are a number of ways to modify the configuration of a request
# * you can add directly to a request
}

